+++ donation-service/src/main/java/com/gohelpfund/api/v1/donation_service/wallets/models/wallet/promise/PromiseWalletBacker.java
@@ -31,7 +31,7 @@ public class PromiseWalletBacker implements Serializable {
 
     @JsonProperty("total_amount")
     @Column(name = "total_amount", nullable = false)
-    private Integer totalAmount;
+    private Double totalAmount;
 
     public PromiseWalletBacker(){
     }
@@ -60,11 +60,11 @@ public class PromiseWalletBacker implements Serializable {
         this.fundraiser_id = fundraiser_id;
     }
 
-    public int getTotalAmount() {
+    public Double getTotalAmount() {
         return totalAmount;
     }
 
-    public void setTotalAmount(int totalAmount) {
+    public void setTotalAmount(Double totalAmount) {
         this.totalAmount = totalAmount;
     }
