+++ donation-service/src/main/java/com/gohelpfund/api/v1/donation_service/wallets/models/wallet/promise/PromiseWalletDetails.java
@@ -25,7 +25,7 @@ public class PromiseWalletDetails implements Serializable {
 
     @JsonProperty("balance")
     @Column(name = "balance")
-    private Integer balance;
+    private Double balance;
 
     @JsonProperty("transactions")
     @Transient
@@ -55,11 +55,11 @@ public class PromiseWalletDetails implements Serializable {
         this.entityId = entityId;
     }
 
-    public Integer getBalance() {
+    public Double getBalance() {
         return balance;
     }
 
-    public void setBalance(Integer balance) {
+    public void setBalance(Double balance) {
         this.balance = balance;
     }
 
@@ -89,7 +89,7 @@ public class PromiseWalletDetails implements Serializable {
         return this;
     }
 
-    public PromiseWalletDetails withBalance(Integer balance){
+    public PromiseWalletDetails withBalance(Double balance){
         this.setBalance(balance);
         return this;
     }
