+++ fundraiser-service/src/main/java/com/gohelpfund/api/v1/fundraiser_service/model/wallet/promise/PromiseWalletTransaction.java
@@ -32,7 +32,7 @@ public class PromiseWalletTransaction implements Serializable {
     private String transactionType;
 
     @JsonProperty("amount")
-    private Integer amount;
+    private Double amount;
 
     @JsonProperty("sender_name")
     private String senderName;
@@ -72,11 +72,11 @@ public class PromiseWalletTransaction implements Serializable {
         this.transactionType = transactionType;
     }
 
-    public Integer getAmount() {
+    public Double getAmount() {
         return amount;
     }
 
-    public void setAmount(Integer amount) {
+    public void setAmount(Double amount) {
         this.amount = amount;
     }
 
@@ -108,7 +108,7 @@ public class PromiseWalletTransaction implements Serializable {
         return this;
     }
 
-    public PromiseWalletTransaction withAmount(Integer amount) {
+    public PromiseWalletTransaction withAmount(Double amount) {
         this.setAmount(amount);
         return this;
     }
